Avoid circular references by merging two specs isolating same specs
Consider the below setup. For the repro query, there was an additional
redistribute motion added while joining zoo and bar on zoo.a = bar.a
even though it was not required.

Why a redistribute is not required?

The join of bar and foo is already distributed by bar.a zoo being
distributed by a, there shouldn't have been a redistribute motion on the
output of join between foo and bar. Because as per equivalence zoo.a =
bar.a = foo.a

```sql
create table foo(a int, b int, primary key(a));
   create table bar(a int, b int, primary key(a));
   create table zoo(a int, b int, primary key(a));
   explain select * from foo left join bar on foo.a=bar.a left join zoo on zoo.a=bar.a;
                                              QUERY PLAN
   -------------------------------------------------------------------------------------------------
    Gather Motion 3:1  (slice2; segments: 3)  (cost=0.00..439.00 rows=3 width=24)
      ->  Nested Loop Left Join  (cost=0.00..439.00 rows=1 width=24)
            Join Filter: true
            ->  Redistribute Motion 3:3  (slice1; segments: 3)  (cost=0.00..437.00 rows=1 width=16) ===> This motion is not required because bar is already distributed on a.
                  Hash Key: bar.a
                  ->  Nested Loop Left Join  (cost=0.00..437.00 rows=1 width=16)
                        Join Filter: true
                        ->  Table Scan on foo  (cost=0.00..431.00 rows=1 width=8)
                        ->  Index Scan using bar_pkey on bar  (cost=0.00..6.00 rows=1 width=8)
                              Index Cond: bar.a = foo.a
            ->  Index Scan using zoo_pkey on zoo  (cost=0.00..2.00 rows=1 width=8)
                  Index Cond: zoo.a = bar.a
    Settings:  optimizer=on
    Optimizer status: PQO version 2.53.1
```

Problem:
The existing code SetHashedEquiv was Creating distribution specs which
have circular references. The intention of the SetHashedEquiv code block
(introduced in commit 7d4bcd0) was to create a distribution spec
considering the equivalent hash spec so that ORCA knows that zoo.a =
bar.a = foo.a. Before commit 7d4bcd0 only the outer distribution spec
was returned and orca didn't know if foo.a is equivalent to zoo.a so
added a motion on the lower nestloop join in above plan.

However, in an attempt to fix that, the code block SetHashedEquiv in
some cases created distribution spec which had circular references.
Consider the below case:
(-> implies equivalent)
Outer Distribution Spec: A->B
Inner Distribution Spec: C->B

In the above case essentially Spec A == B == C. But the way
SetHashedEquiv created the new spec was to compare the Inner
distribution spec C and B with the outer distribution spec A only. So,
since Inner Spec C != Outer Spec A and Inner Spec B not equivalent to
Spec A as inner spec does not know if B == A, the returned spec had the
outer distribution spec append to the inner distribution spec, i.e
C->B->A->B. However if you notice now, this forms a infinite circular
loop, so it is of the form C->B->A->B->A->B->A and so on causing a
crash.

The object CDistributionSpecHash has an member variable m_pdshashedEquiv
which either points to NULL or the equivalent Hash spec. So essentially
this object forms a chain of all the equivalent hash specs using the
m_pdshashedEquiv variable.

This PR fixes the issue by merging the two spec in a better way. The
resulting spec will only contain A->C->B and the plan will not have a
redistribute motion for the repro query.