Fix MarkedUsedEdges()
MarkedUsedEdges() compared the `edge` with each of the predicate in the
component's predicates. This worked fine when edge consisted of a single
predicate like `t1.a = t2.b`. But if the edge had more than a single
predicate like `t1.a = t2.b and t1.c = t1.d` the current erroneously did
not mark it as used.  We fix that in here making sure each of the
predicate of the edge is present in component's predicates.