Remove Master/Standby SyncRepWait Greenplum hack
When the standby master is unavailable, the master will not block on commits
even though we enable synchronous replication. This is because we have a
Greenplum hack which checks if the WAL stream with the standby master is
valid. If the stream is invalid, the master will quickly skip the SyncRepWait
and continue on its commit.

Remove this hack in order to make Master/Standby and Primary/Mirror WAL
replication more similar.

Co-authored-by: Shaoqi Bai <sbai@pivotal.io>