Add total number of bindings and alternatives generated info
For a group expression, the pattern defined by a xform is extracted, and the
xform can be applied to all the possible bindings for that group expression.
When the GUC optimizer_print_optimization_stats is on, the number of times the
xform is called is printed, however, even though there were several bindings
and the xform was applied several times, we record the number of calls to that
xform as 1. This commit adds more debugging information by printing the total
number of bindings and the total number of alternatives generated for a given
group expression.

Previous output:
CXformInnerJoin2IndexGetApply: 1 calls, 10 ms

Now:
CXformInnerJoin2IndexGetApply: 1 calls, 2 total binding, 2 alternatives generated, 10ms

This helps in identifying the total number of patterns which get extracted for
the given group expression which can often be a factor for long running
queries.

Signed-off-by: Ashuka Xue <axue@pivotal.io>