Improve analyzedb performance on partitioned tables
Until now, when we issued an "analyze <partition>" command, we would
produce new root-level statistics (merge) for each such command,
assuming we had valid HLL stats for every partition and every column.
That is very expensive, and could take hours to complete, if we
need to update many partitions. The cost is roughly

u * p * c

where u is the number of incrementally updated partitions, p is the
total number of partitions and c is the number of columns.

This commit adds a new guc, optimizer_analyze_enable_merge_of_leaf_stats
that enables creating new root-level stats when we analyze an
individual partition. The guc is set to "on" by default, so there is
no change in the default behavior.

The analyzedb program, however, will turn this guc off and therefore
suppress generating the root stats for every partition. The root-level
stats will be generated when analyzedb issues the final
"analyze rootpartition" command. For tables with very many partitions,
this can speed up the process by several orders of magnitude when we
do a full analyze, the new cost is now p * c.

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>