Add 2 new xforms for index apply
The new xforms use the new CPatterNode for a patterns of the form

CPatternNode(Leaf, Tree, Tree)

The CPatternNode matches both an inner and a left outer join.
The Tree for the right child can contain arbitrary operators.
To avoid an explosion of the search space, we add two conditions
to the xform: First, the xform is applied only once, not for all
bindings of the Tree node. Second, the xform is applied only if
the right child has a "join depth" of 1, excluding any children
that are complex and wouldn't satisfy the conditions of this xform
anyway.

Co-authored-by: David Kimura <dkimura@vmware.com>
Co-authored-by: Hans Zeller <hzeller@vmware.com>