Support for project nodes in index nested loop joins
When generating required distribution specs for its children,
CPhysicalInnerIndexNLJoin will start with its inner child and send it
an ANY required distribution spec. It will then force the outer child
to match the inner's distribution spec (or require a broadcast on the outer).

Now, assume we have CPhysicalComputeScalar as the inner child. This
node, in CPhysicalComputeScalar::PdsRequired will currently require
its child to be REPLICATED (opt request 1) or SINGLETON (opt request
0), if the expression has any outer references. This won't work, since
the underlying table has neither of these distribution schemes and
since we don't want any motions between the index join and the
index get.

This commit changes the behavior of a CPhysicalComputeScalar. If it
senses that it is part of an index nested loop join, it will just
propagate the required distribution spec from the parent.
How does it sense that? By the required ANY distribution spec that
allows outer references. This request is generated in only two places:
CPhysicalInnerIndexNLJoin::PdsRequired and
CPhysicalLeftOuterIndexNLJoin::PdsRequired, so it is only used in the
context of an index NLJ. This behavior is similar to what the CPhysicalFilter
node does, the other node allowed between an index NLJ and the get.