Handle single outer references in subquery project lists
In a previous commit, we added code to fall back gracefully when a
subquery select list contains a single outer ref that is not part
of an expression, such as in

select * from foo where a is null or a = (select foo.b from bar)

Now, we added a fix that will allow us to handle such queries in ORCA,
by adding a project in the translator that will echo the outer ref
from within the subquery, and using that projected value in the
select list of the subquery. This ensures that we use a NULL value
for the outer ref when the subquery returns no rows.

We also had to fix an assert in the code that combines multiple
projects into one - we need to allow for the case described above.

Co-authored-by: Shreedhar Hardikar <shardikar@vmware.com>