Project certain outrrefs in the targetlist of subqueries
In a previous commit (df5f06d), we added code to fall back gracefully
when a subquery select list contains a single outer ref that is not part
of an expression, such as in

select * from foo where a is null or a = (select foo.b from bar)

This commit adds a fix that allows us to handle such queries in ORCA
by adding a project in the translator that will echo the outer ref
from within the subquery, and using that projected value in the
select list of the subquery. This ensures that we use a NULL value
for the outer ref when the subquery returns no rows.

Also note that this is still skipped for grouping cols in the target
list. This was done to avoid regression for certain queries, such as:

select *
from A
where not exists (select sum(C.i)
                  from C
                  where C.i = A.i
                  group by a.i);

ORCA is currently unable to decorrelate sub-queries that contain project
nodes, So, a `SELECT 1` in the subquery would also cause this
regression. In the above query, ORCA adds `a.i` to the target list of
the subquery, that gets an echo projection (as described above), and
thus cannot be decorelated by ORCA. In this case, we decided to maintain
existing behavior until ORCA is able to handle projections in subqueries
better.