Address review comments and give bitmap scan an initial advantage
A ctest failure reminded me that when Ashuka worked on this initially,
we used a smaller constant init cost (dInitScan in the code), to give
index scans a small initial advantage. Otherwise, small queries will
have more or less random plans, all of which cost around 431, the value
of the initial cost. Added a 10% advantage of the bitmap scan and
fixed ctest failures.

Note that this doesn't make a big difference in the calibration tests,
which are run with larger tables and therefore with costs that are
quite a bit larger than 431.