Pass outer ref constraints to several methods
To generate inferred predicates for outer refs, we need to pass the
constraints for outer refs to methods
and CExpressionPreprocessor::PexprFromConstraints(),
CExpressionPreprocessor::PexprFromConstraintsScalar(), and
CExpressionPreprocessor::PexprScalarPredicates().

Initially NULL, the new argument is set to the constraints of the
current expression when we enter a subquery through a call to
CExpressionPreprocessor::PexprFromConstraintsScalar().