Infer predicates from constraints on outer refs
When generating inferred predicates, combine the local equivalence class
for the colref with that of the outer ref at the outer scope. With this
combined equivalence class, search both the current expression's constraints
as those of the outer refs for predicates to infer.

This will help queries such as

select *
from   foo
where  foo.a in (select max(bar.a) from bar where foo.b=bar.b)
   and foo.b < 2;

We can now infer a predicate bar.b < 2 inside the subquery. The outer scope
contributes the constraint foo.b < 2 and the inner scope contributes the
equivalence class { foo.b, bar.b }.