[Expr->DXL] Map columns from the partitioned table to each child partition
This uses CLogicalDynamicGet::m_root_col_mapping_per_part, which is a
map from root colref -> col index in child partition tabledesc, per
child partition in m_partition_mdid.

The mapping is constructed by comparing the column names in the root
table to those in each sub-partition. Since each sub-partition may
re-order the columns and/or have dropped columns, this mapping is
necessary to properly convert columns between the root & child
partitions column "namespaces" correctly.

This mapping is then used in CExpressionPreprocessor for static
partition elimination and in the Expr->DXL translation for translating
the project list & filter per child partition.

Also, ignore dropped columns while translating PartConstraints so that
this mapping can be used consistently for PartConstraints translation as
well as Expr->DXL translation. This is needed because the table
descriptors for leaf partition tables don't contain dropped columns and
they are not really needed for translating PartConstraints anyway, since
dropped columns cannot participate in a partition constraint.
(This is not going to different than CheckConstraints, which currently
uses dropped columns. Added a FIXME to correct this since dropped
columns are not needed there for the same reasons)

Note also that the *first* child of the Append retains the old colids,
and we create new ones for every subsequent partition. That way, there
are no holes in the colid space once translated!