Limit INDF predicate selectivity to 100%
In the case where we joined two histograms that consisted entirely of
NULL values we ended up with a row count estimate that was larger than
the cartesian product of the two tables, which is impossible to achieve
with a join.

This triggered the assert GPOS_ASSERT(CStatistics::MinRows <= scale_factor)
in CJoinStatsProcessor::CalcJoinCardinality().