MDP changes
Took about 80% of the MDPs that used the old default optimizer_join_order
'exhaustive' and converted them to 'exhaustive2' by altering the traceflags
in the same way a guc change would have. As expected, this lead to some
plan changes, mostly join order changes and different aggregation and
motions. MDPs that were found to have more serious issues are handled in
separate commits.