New class to track common subgroups in xform results
Added a new class, CXformResultSubGroup, that can be used to track
common subgroups in xform results, so they can immediately be put into
the same MEMO groups, without having to rely on other xforms to discover
equivalent groups and perform a group merge.

This new class can be subclassed by individual xforms (e.g.
CXformExpandNAryJoinDPv2), to contain additional info needed by the xform.

- An xform generates multiple alternatives with subtrees that are semantically
  equivalent. For example, it might generate two alternatives a1 and a2:
    a1: ((t1 join t2) join t3) join t4
    a2: ((t1 join t3) join t2) join t4
  not only are a1 and a2 equivalent (as all alternatives have to be),
  the subtrees that join tables { t1, t2, t3 } are also semantically equivalent
  and the xform is aware of this equivalence.

- The xform provides an optional hash table in the CXformResult object
  that maps both expressions ((t1 join t2) join t3) and ((t1 join t3) join t2)
  to the same CXformResultSubGroup object. Note that at this point the MEMO
  group for these expressions doesn't exist yet, we only know that we want
  both expressions to share the same MEMO group.

Some notes:

The CXformResult object with the subgroup info survives beyond the
CJoinOrderDPv2 object that contains all these expressions and
CXformResultSubGroup objects. The CXformResult holds on to a refcount
on those objects, to prevent deleting them with the CJoinOrderDPv2
object.

Sometimes we use the exact same CExpression in multiple
alternatives. In that case we don't need this commit, since these
expressions should hash to the same value and therefore be inserted
only once into CMemo.

The CXformResult object can be used for multiple bindings and therefore
multiple calls of the CXform::Transform method. We currently don't consider
common subgroups across bindings and this should not be an issue for the
CXformExpandNAryJoinDPv2 xform.