Make normalizer lazy
The PexprNormalize method is used in various places, mainly in the
preprocessor, but also for the alternatives generated by xforms.

This commit makes the normalizer somewhat "lazy", meaning that it may
return the original expression and not a copy when normalization didn't
result in any changes. Note that there isn't a guarantee that this happens,
in some less common cases we may return a copy even if there are no changes.

The main reason for this commit is that in the DPv2 xform, we need to know
whether the normalizer changes the generated alternatives. A side effect of
the commit might be a slight performance improvement.