Consumer side for common subgroups
Until now, we just provided the information for common subgroups. Now,
we use this info to put equivalent expressions into the same CGroup.

To have the necessary info available, we pass the CXformResult object
to a few more methods.

When we insert the first (and maybe only) expression of a subgroup
into CMemo, we remember the new CGroup in the CXformResultSubGroup
object associated with this (and maybe other) CExpression(s).

If we come across another CExpression for the same subgroup, then
we know into which CGroup we should insert the equivalent expression.
This reduces the overall number of groups and allows for additional
choices during the optimization phase. It might also reduce the
number of group merges, although in my testing I didn't find that
to be the case.