Add support for internal evaluation of date expressions in MDP files
Since we don't have GPDB available in the gporca_test program, we didn't
try to evaluate expressions on types other than integer types. For integer
types, we just used the stats values (LINT values) to evaluate comparison
operators.

With this commit, we can do the same on date columns, which have a DOUBLE
value that represents the number of microseconds since 2000-01-01
(older dates have negative values). We could later add timestamp values
and probably time values as well, if we want.

For now, the main goal is to get 22 MDPs that use date partitioning
columns to work.