Disallow btree index scan on AO table
Method CXformUtils::FIndexApplicable() allowed a btree index to be
used for an IndexGet on an AO table, which is not supported.

Fixing the method to return false for this situation.

Now, the CXformJoin2IndexGetApply xform should no longer generate
an alternative for a btree index on an AO table.

This is a backport of master PR https://github.com/greenplum-db/gpdb/pull/11399.