Limit join scale factor to cardinality of larger table
When we treat multiple join predicates as independent, limit
the resulting scale factor to the cardinality of the larger table.
This has the effect of a join condition on a unique key with a
referential integrity constraint. In other words, the estimated
join cardinality is the cardinality of the larger of the two tables.

The general formula for the cardinality of an expression

  R join S on R.c = S.c

is

  |R join S on R.c = S.c| = |R| * |S| / max(NDV(R.c), NDV(S.c))

|x| stands for the cardinality estimate for x.

The expression in the denominator, max(NDV(R.c), NDV(S.c)), is
the scale factor.

For a single equi-join predicate like in this example, the
scale factor can't exceed max(|R|, |S|), since the NDVs can't
be greater than the cardinality of the table.

This commit applies a similar rule when we have multiple scale
factors. For example, look at this join:

  R join S on R.a = S.a and R.b = S.b

If we assume columns a and b are not correlated, then we'll just
multiply the selectivities:

  |R| * |S| / (max(NDV(R.a), NDV(S.a)) * max(NDV(R.b), NDV(S.b)))

In this formula, we want to make sure that the scale factor
max(NDV(R.a), NDV(S.a)) * max(NDV(R.b), NDV(S.b)))
doesn't exceed max (|R|, |S|).

In other words, we assume that the join predicate

  R.a = S.a and R.b = S.b

selects 1 or more rows from the larger table, given a row from the
smaller table. We make the same assumption when we only have a
single join predicate like in the first example, that's the
assumption the formula for the cardinality is based on.