Fix SIGSEGV/assert violation with certain functions with no arguments
When testing whether a function (e.g random() or timeofday()) can be
used as a monotonically increasing function in stats for partition
columns, we tried to access the function argument without checking for
its arity which can lead to an illegal access. This commit fixes this
issue by adding the check. Other functions with arity 0 are evaluated
at planning time and don't run into this problem. Examples: pi(),
current_timestamp, current_schema().