Change the partition elimination expr for IS [NOT] NULL
Changing this expression for range-partitioned tables. I didn't
understand the existing logic. As far as I can tell, the NULL value
can only go into the default range partition, but the default
partition can also have non-null values (we probably don't want to
write code that detects when this is not possible).

Therefore, an IS NULL predicate selects the default range partition
while an IS NOT NULL predicate selects all range partitions.