Use NDV to calculate CBucket::MakeBucketSingleton frequency
This is a port of https://github.com/greenplum-db/gpdb/pull/11507 to master.
Cherry-picked 7dbc559607af660a804453b0035a6820d73ab924 from 6X_STABLE branch
and regenerated plan diffs.

Issue with bucket width approach is that when width to NDV ratio is
high, meaning the average distance between adjacent values is also high,
then the ratio of a singleton point is low and can lead to
underestimates of the singleton bucket frequency for point queries where
we often expect that the value exists in the bucket.

* Fixup tests and add testcase

* Use NDV in CBucket::GetOverlapPercentage for singleton

* Revert "Remove Orca assertions when merging buckets"
   This reverts commit 34ae3d9.

Co-authored-by: David Kimura <dkimura@vmware.com>