Clean stale state directories in analyzedb
This is a port of 6X PR https://github.com/greenplum-db/gpdb/pull/11654.

The analyzedb utility leaves a state directory each time it is run (in
a way that it actually performs ANALYZE commands). These state
directories are stored in a subdirectory db_analyze below the master
data directory.

With this fix we preserve only state directories that less than 8 days
old, with one exception: We also preserve the 3 most recent state
directories regardless of age, plus the current state directory.
All others are removed automatically.

Users should not need to deal with state directories, except when
using the --clean_last option that restores the state of the previous
analyzedb run. Leaving at least 3 historic state directories allows to
use this option at least 3 times in a row (more often if the state
directories are less than 8 days old).