Fix the includecheck build failure on MacOS for GPDB 5X
Apple, in its infinite wisdom, decided that MacOS should not have
a /usr/include directory anymore. The src/test/regress/checkinc.py
program relies on that directory, however, causing failures in
some build steps.

The fix checks for the new location of the usr/include files and tests
it in addition to /usr/include. This applies to MacOS Catalina and
later versions.