Fix minirepro bug with column types that are not built-in
In the generated SQL, minirepro used an unqualified name for the
type, which didn't work for types like citext that usually come
from a schema other than pg_catalog and other than the schema of
the table to be dumped.

The fix is to use a qualified name. For example, citext becomes
something like "public"."citext".