gpsd and minirepro fixes from 6X
This is a port from 6X PR https://github.com/greenplum-db/gpdb/pull/11748

Fix gpsd and minirepro bug with column types that are not built-in

In the generated SQL, gpsd and minirepro used an unqualified name for
the type, which didn't work for types like citext that usually come
from a schema other than pg_catalog and other than the schema of the
table to be dumped.

The fix is to use a qualified name. For example, citext becomes
something like "public"."citext".

Fix gpsd and minirepro bug with embedded double quotes and backslashes

When values in pg_statistic.stavalues<n> had embedded double quotes or
backslashes, gpsd and minirepro generated a single set of escape
characters (backslash) but the correct syntax for an embedded double
quote in an array constant is \\", because it goes through two layers
of escape processing. For embedded backslashes we need to use \\\\.